requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.1.2
fonttools==4.33.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy
openai==0.20.0
openpyxl==3.0.10
packaging==21.3
pandas
pandas-stubs
Pillow==9.1.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
tqdm==4.64.0
twilio==7.9.3
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
